+++ launch/dynamo-run/src/lib.rs
@@ -67,6 +67,8 @@ pub async fn run(
     //
     // Create an engine
     //
+    tracing::info!("********* before output validation DEBUG **********");
+
 
     let out_opt = out_opt.unwrap_or_else(|| default_engine_for(&local_model));
     print_cuda(&out_opt);
@@ -83,7 +85,7 @@ pub async fn run(
         rt.clone(),
     )
     .await?;
-
+    tracing::info!("********* DEBUG **********");
     //
     // Run in from an input
     //
