+++ lib/llm/src/entrypoint/input/grpc.rs
@@ -7,7 +7,7 @@ use crate::{
     discovery::{ModelManager, ModelWatcher, MODEL_ROOT_PATH},
     engines::StreamingEngineAdapter,
     entrypoint::{self, input::common, EngineConfig},
-    http::service::service_v2,
+    grpc::service::kserve,
     kv_router::KvRouterConfig,
     types::openai::{
         chat_completions::{NvCreateChatCompletionRequest, NvCreateChatCompletionStreamResponse},
@@ -20,27 +20,25 @@ use dynamo_runtime::{DistributedRuntime, Runtime};
 
 /// Build and run an HTTP service
 pub async fn run(runtime: Runtime, engine_config: EngineConfig) -> anyhow::Result<()> {
-    let mut http_service_builder = service_v2::HttpService::builder()
-        .port(engine_config.local_model().http_port())
-        .enable_chat_endpoints(true)
-        .enable_cmpl_endpoints(true)
-        .enable_embeddings_endpoints(true)
+    tracing::info!("********* DEBUG **********");
+    let mut grpc_service_builder = kserve::GrpcService::builder()
+        .port(engine_config.local_model().http_port()) // [WIP] generalize port..
         .with_request_template(engine_config.local_model().request_template());
 
-    let http_service = match engine_config {
+    let grpc_service = match engine_config {
         EngineConfig::Dynamic(_) => {
             let distributed_runtime = DistributedRuntime::from_settings(runtime.clone()).await?;
             let etcd_client = distributed_runtime.etcd_client();
             // This allows the /health endpoint to query etcd for active instances
-            http_service_builder = http_service_builder.with_etcd_client(etcd_client.clone());
-            let http_service = http_service_builder.build()?;
+            grpc_service_builder = grpc_service_builder.with_etcd_client(etcd_client.clone());
+            let grpc_service = grpc_service_builder.build()?;
             match etcd_client {
                 Some(ref etcd_client) => {
                     let router_config = engine_config.local_model().router_config();
                     // Listen for models registering themselves in etcd, add them to HTTP service
                     run_watcher(
                         distributed_runtime,
-                        http_service.state().manager_clone(),
+                        grpc_service.state().manager_clone(),
                         etcd_client.clone(),
                         MODEL_ROOT_PATH,
                         router_config.router_mode,
@@ -52,7 +50,7 @@ pub async fn run(runtime: Runtime, engine_config: EngineConfig) -> anyhow::Resul
                     // Static endpoints don't need discovery
                 }
             }
-            http_service
+            grpc_service
         }
         EngineConfig::StaticRemote(local_model) => {
             let card = local_model.card();
@@ -60,8 +58,8 @@ pub async fn run(runtime: Runtime, engine_config: EngineConfig) -> anyhow::Resul
 
             let dst_config = DistributedConfig::from_settings(true); // true means static
             let distributed_runtime = DistributedRuntime::new(runtime.clone(), dst_config).await?;
-            let http_service = http_service_builder.build()?;
-            let manager = http_service.model_manager();
+            let grpc_service = grpc_service_builder.build()?;
+            let manager = grpc_service.model_manager();
 
             let endpoint_id = local_model.endpoint_id();
             let component = distributed_runtime
@@ -98,23 +96,23 @@ pub async fn run(runtime: Runtime, engine_config: EngineConfig) -> anyhow::Resul
             .await?;
             manager.add_completions_model(local_model.display_name(), completions_engine)?;
 
-            http_service
+            grpc_service
         }
         EngineConfig::StaticFull { engine, model, .. } => {
-            let http_service = http_service_builder.build()?;
+            let grpc_service = grpc_service_builder.build()?;
             let engine = Arc::new(StreamingEngineAdapter::new(engine));
-            let manager = http_service.model_manager();
+            let manager = grpc_service.model_manager();
             manager.add_completions_model(model.service_name(), engine.clone())?;
             manager.add_chat_completions_model(model.service_name(), engine)?;
-            http_service
+            grpc_service
         }
         EngineConfig::StaticCore {
             engine: inner_engine,
             model,
             ..
         } => {
-            let http_service = http_service_builder.build()?;
-            let manager = http_service.model_manager();
+            let grpc_service = grpc_service_builder.build()?;
+            let manager = grpc_service.model_manager();
 
             let chat_pipeline = common::build_pipeline::<
                 NvCreateChatCompletionRequest,
@@ -129,18 +127,18 @@ pub async fn run(runtime: Runtime, engine_config: EngineConfig) -> anyhow::Resul
             >(model.card(), inner_engine)
             .await?;
             manager.add_completions_model(model.service_name(), cmpl_pipeline)?;
-            http_service
+            grpc_service
         }
     };
-    tracing::debug!(
-        "Supported routes: {:?}",
-        http_service
-            .route_docs()
-            .iter()
-            .map(|rd| rd.to_string())
-            .collect::<Vec<String>>()
-    );
-    http_service.run(runtime.primary_token()).await?;
+    // tracing::debug!(
+    //     "Supported routes: {:?}",
+    //     grpc_service
+    //         .route_docs()
+    //         .iter()
+    //         .map(|rd| rd.to_string())
+    //         .collect::<Vec<String>>()
+    // );
+    grpc_service.run(runtime.primary_token()).await?;
     runtime.shutdown(); // Cancel primary token
     Ok(())
 }
