+++ lib/llm/src/lib.rs
@@ -18,6 +18,7 @@ pub mod endpoint_type;
 pub mod engines;
 pub mod entrypoint;
 pub mod gguf;
+pub mod grpc;
 pub mod http;
 pub mod hub;
 // pub mod key_value_store;
